# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Visitor Management System API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/appointment": {
      "get": {
        "summary": "Get all appointments",
        "tags": [
          "Appointment"
        ],
        "responses": {
          "200": {
            "description": "List of all appointments"
          }
        }
      },
      "post": {
        "summary": "Create a new appointment",
        "tags": [
          "Appointment"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "visitorId": {
                    "type": "integer"
                  },
                  "purpose": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "time": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "integer"
                  },
                  "department": {
                    "type": "string"
                  },
                  "host": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created successfully"
          }
        }
      }
    },
...
    "/api/v1/appointment/{appointment_id}": {
      "get": {
        "summary": "Get an appointment by ID",
        "tags": [
          "Appointment"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Appointment ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment details"
          },
          "404": {
            "description": "Appointment not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing appointment",
        "tags": [
          "Appointment"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Appointment ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "visitorId": {
                    "type": "integer"
                  },
                  "purpose": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "time": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "integer"
                  },
                  "department": {
                    "type": "string"
                  },
                  "host": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete an appointment",
        "tags": [
          "Appointment"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Appointment ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment deleted successfully"
          },
          "404": {
            "description": "Appointment not found"
          }
        }
      }
    },
    "/api/v1/auth/send-otp": {
      "post": {
        "summary": "Send OTP to visitor",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully"
          }
        }
      }
    },
    "/api/v1/auth/verify-otp": {
      "post": {
        "summary": "Verify OTP for visitor",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "otp": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully"
          }
        }
      }
    },
...
    "/api/v1/auth/reset-password/{token}": {
      "post": {
        "summary": "Reset password",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset successfully"
          }
        }
      }
    },
    "/api/v1/auth/check-phone": {
      "post": {
        "summary": "Check if a phone number exists",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Phone number exists"
          },
          "404": {
            "description": "Phone number not registered"
          }
        }
      }
    },
    "/api/v1/auth/login-phone": {
      "post": {
        "summary": "Login using phone number",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "404": {
            "description": "Phone number not registered"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login using email and password",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid email or password"
          }
        }
      }
    },
...
  },
...
}